Update main.go
Added minSdkVersion modification to Android manifest changer script.

Enhanced the Android manifest changing script to enable modification of the 'minSdkVersion' attribute. This enhancement addresses Google Play's requirement for apps to target an API level close to the latest Android release. With this update, the script now assists any developer in meeting the August 31, 2023, deadline to target Android 13 (API level 33), facilitating compliance for a broader range of Android apps on the Google Play Store.